Move conversions.py to omero
A recursive import seems to have prevented
bin/omero from operating. Moving to the top
level omero/ package seems to have fixed it.
