+++ components/tools/OmeroPy/src/omero/conversions.py
@@ -32,7 +32,7 @@ class Conversion(object):
     strings are placed directly into code. If the proper imports are in place,
     then a top-level Conversion (usually of type Add or
     Mul is returned from the evaluation).
-    """)
+    """
 
     def __init__(self, *conversions):
         self.conversions = conversions
